[7184] - HesaMapper module namespace
